mx7dvar_som: Make SPL DDR configuration match the DCD table
As the non-SPL code has been tested for a long time and has been proven to be
reliable, let's configure the DDR in the exact same way as the non-SPL case.

The idea is simple: just use the DCD table and write directly to the DDR
registers.

Retrieved the DCD table from:
board/variscite/mx7dvar_som/imximage.cfg

This will also provide the most flexibility when we'll use the EEPROM to
store the DDR configuration.
And another benefit is that the SPL size is reduced.

Signed-off-by: Eran Matityahu <eran.m@variscite.com>