mx7dvar_som: Enable SDP support
Enable Serial Download Protocol (SDP) support in SPL and U-Boot.
This is useful to make use of the imx_usb_loader tool to
download both SPL and then the U-Boot image. The sdp U-Boot
command allows to enumerate as SDP capable device again, e.g. to
download a Linux kernel and/or U-Boot script.

See doc/README.sdp for more info.

Apply the above to the following defconfigs:
mx7dvar_som_defconfig

The default boot command when booting from USB is bootcmd_mfg.
It can be defined in the in the boards default env (we leave it
empty by default).

Signed-off-by: Eran Matityahu <eran.m@variscite.com>