MA-16095 set evk_8mm/evk_8mn usb dr_mode to "otg"
When trying to remove "CONFIG_USB_TCPC=y" to avoid ramdom issue of
synchronous abort when enter fastboot, evk_8mm and evk_8mn can't
enter fastboot mode, it's because init_type of the usb udevice private
data is host if tcpc configuration is removed, if this init_type is
host/peripheral, then the usb is forced to be this mode, can not be
initialized as another.

init_type of the usb udevice private data is firstly set based on the
"dr_mode" in dtb, if "dr_mode" is not set, it will be set by other ways.
here set the "dr_mode" property for the usb device to be "otg", then the
device can be configured as either host or peripheral.

Change-Id: Idadab7bb07e74bfc1f8768490c07127a35a02572
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>