imx8qxp_var_som: Fix DRAM size detection on old SOM revisions
On early VAR-SOM-MX8X SOM revisions the EEPROM is not connected
to SCU, so reading its contents in SFCW and passing it to u-boot
is not possible. On these SOMs the DRAM size is always 2GB. The
DRAM size detection code relied on DEFAULT_SDRAM_SIZE being 2GB
to properly detect DRAM size on these SOMs. This assumption was
broken when support for 1GB DRAM size was added recently.

Fix the problem by setting DEFAULT_SDRAM_SIZE back to 2GB.