MA-19357 imx8mq: enhance security for android
Setting the CSU configs to enhance the imx8mq security. And as
the ocram space reserved for TF-A is very limited (64 KB), the
code size would overflow the limit if setting csu configs in TF-A,
so this commit will set it in spl.

Signed-off-by: Ji Luo <ji.luo@nxp.com>
Reviewed-by: Ye Li <ye.li@nxp.com>
Change-Id: If9d2c20401cb256174aa2e9a72cae2686e58c7bd
(cherry picked from commit 9f08e16a291e4d42b6ccb7386b506f6317931bb5)