MA-19663 Configure boot security patch level
The boot security patch level should be set in bootloader
and it will be returned in the TEE enforced authorization
list.

After building with the "BOOT_SECURITY_PATCH" set, the boot
security patch level would be set as a vbmeta property and
would be saved in the vbmeta struct in boot image.

This commit would parse the boot security patch level (which
is in YYYY-MM-DD format) and pass it to keymaster TA.

Test: Set and get boot security patch level &&
      VtsHalKeymasterV4_0Target module run

Change-Id: Iac75579c680e8b80b28d3a82cd17cbcd8a93509e
Signed-off-by: Ji Luo <ji.luo@nxp.com>
(cherry picked from commit 8816d0a5327fc5dc1e2460db8185d39e5b2dd9ce)