imx8qxp_var_som: enable carrier board phy
Enable fec2, since it is now working with the recent patches:
- net: dm: fec: configure phy-supply-vselect regulator for fec2
- dm: fec: add phy-supply-vselect regulator
- imx8qxp_var_som: add define for CONFIG_FEC_MXC_MDIO_BASE

Also, starting in Symphony v1.6, the pull-up resistor on the ethernet phy
reset line is not populated. In Linux, when fec1 is probed, the mii
initialization tries reading both fec1 and fec2 phy-ids prior to releasing
the reset line on fec2 phy. This causes fec2 to fail. See:

https://github.com/varigit/linux-imx/blob/14b26bad1d957d527424f8bdb1f605f62e593943/drivers/net/ethernet/freescale/fec_main.c#L2080-L2105

Enabling fec2 in U-Boot will assure the reset line, controlled by the
pca9534, will be released for Linux.

Signed-off-by: Nate Drude <nate.d@variscite.com>