imx8qm_var_som: enable Symphony board phy
Enable fec2, since it is now working with the recent patches:
- imx8qm_var_som: add eth1 regulators
- imx8qm_var_som: fix i2c1 pinmux
- net: dm: fec: add phy-supply-vselect regulator

Also, starting in Symphony v1.6, the pull-up resistor on the ethernet phy
reset line is not populated. In Linux, when fec1 is probed, the mii
initialization tries reading both fec1 and fec2 phy-ids prior to releasing
the reset line on fec2 phy. This causes fec2 to fail. See:

https://github.com/varigit/linux-imx/blob/1b88e6a2b5d51f17cbd24dd744ee86e1dfeb081f/drivers/net/ethernet/freescale/fec_main.c#L2172-L2196

Enabling fec2 in U-Boot will assure the reset line, controlled by the
pca9534, will be released for Linux.

Derived from:
https://github.com/varigit/uboot-imx/commit/a34510e80d8f98a30fa41100a20d49b68dcc01ec

Signed-off-by: FrancescoFerraro <francesco.f@variscite.com>