MA-20419 Move memory region rdc configure to TEE
The widevine L1 need to support VP9 4K format video,
and the size of secure heap is not enough for it, so need
to expand the size of secure heap.

The restrictions on the secure heap in the spl stage,
and expnasion of secure heap may affect the use of memory in the
uboot stage. In order to prevent affecting uboot from using secure heap
range, the memory policy are not configured in the spl stage by
default. Only widevine TA or confirmationUI TA is loaded, will
configure it in TEE.

Change-Id: Ib8f0a10b26e3900c66757f7a684270a061e62f46
Signed-off-by: zhai.he <zhai.he@nxp.com>