imx8qm_var_som: remove eth1_phy_rst_hog
The reset gpios for the Symphony phy is controlled by the fec driver.
If gpio-hog is using the gpio, the fec driver request for the gpio
fails returning "Device or resource busy". At this point the reset_delay
and reset_post_delay are not parsified and are equal to zero.

Signed-off-by: FrancescoFerraro <francesco.f@variscite.com>