imx93_var_som: dts: Correct USDHC CLK and STROBE pad config settings
Currently, the CLK and STROBE pads for USDHC are configured as a mixture
of pull up and pull down settings. This is incorrect, as these should be
configured as pull downs.

Without this, when using some eMMC chips like Samsung and Micron, CMD8
will always return a CRC error when in HS400/HS400ES modes.

This patch is based on an NXP patch for imx93-11x11-evk which was made
to mainline Linux in commit 62f0147fd4d86620853bee027800f988d3013656.

Signed-off-by: Leonid Segal <leonids@variscite.com>
Signed-off-by: Ken Sloat <ken.s@variscite.com>