imx93_var_som: dts: Disable card detect pin for SD card
In order to provide flexibility for customers who may choose different
pin configurations, disable card detect on the SD card peripheral.

There are multiple ways to achieve this in U-Boot, but the chosen
method using `non-removable` is often used my eMMCs and will simply
allow the drivers to always see the card indicated as present.

Signed-off-by: Leonid Segal <leonids@variscite.com>
Signed-off-by: Ken Sloat <ken.s@variscite.com>