imx93_var_som: Change MMC env offset to 0x700000
Currently, the environment offset is set to 0x400000 in the MMC device.
Unless the user programs the associated fuses, this offset is the
default secondary boot image offset used by the i.MX 93 ROM bootloader.
With certain combinations of environmental variables, the CRC and
beginning of the environment can potentially appear as a valid unsigned
boot image header with a "sw_version" field that is greater than the
primary image's. In this situation, the ROM bootloader will skip the primary
image and attempt to boot using the environment data and hang.

To fix this, move the environment from 0x400000 to 0x700000 in order to
reserve 3 MB at 0x400000 for any actual secondary user image.

Signed-off-by: Ken Sloat <ken.s@variscite.com>