MA-21194 set the rng-seed prop to add entropy to crng entropy pool
The kernel crng driver has an entropy pool, which needs 256 bits entropy
before block access can going on.

The crng driver mainly gets random data from interrupt and timer to mix
into the entropy pool and increases the entropy bits.

With 6.1 kernel, timer is checked with more strict rules and the timer
on i.mx 8ulp is not used to generate entropy. This makes the crng driver
take a long time to be fully initialized and it block the boot up of
Android.

This commit try to get the /chosen/rng-seed property value length and
generate random number to overwrite it. The value will be mixed into
crng entropy pool and provide (8 * lengh_of_the value) bits entropy.

Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Change-Id: I988e51ec441e3f5a8255370ba5e45e39f7e5c08b