LFU-545 imx8ulp: Enable PS13 to workaround M33 PXP/EPDC issue
When PS13 is changed to default disable, M33 meets PXP WFE_A
issue that WFE_A_STORE_IRQ can't be polled, but dither has completed.
So far the only way to make PXP/EPDC working in M33 is power on
the PS13 in SPL.

So we add a workaround, for 8ULP D7 part, SPL will power on the PS13
but for other parts which fused EPDC/PXP don't power on it.

Signed-off-by: Ye Li <ye.li@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>