phy: mxl-8611x: add driver for MaxLinear mxl-8611x PHYs
The MxL86110 is a low power Ethernet PHY transceiver integrated
circuit following the IEEE 802.3 [1] standard. It offers a
cost-optimized solution that is well-suited for routers,
switches, and home gateways. It performs data transmission on
an Ethernet twisted pair copper cable of category CAT5e or higher.
The MxL86110 supports 1000, 100, and 10 Mbit/s data rates.

The current driver implementation supports:
- configuring rgmii from the device tree
- configuring the LEDs from the device tree
- reading extended registers using the mdio command, e.g.:
    => mdio rx ethernet@428a0000 0xa00c
    Reading from bus ethernet@428a0000
    PHY at address 0:
    40972 - 0x2600

Signed-off-by: Nate Drude <nate.d@variscite.com>