variscite: imx8mp-var-dart: Kconfig: remove source freescale/common/Kconfig
The way that we use this file currently means that we have to guard it
in every platform Kconfig.  But it is also required in all NXP
platforms, including non-reference platforms.  Make all options in it
have appropriate dependencies so that we can include it a single time
under arch/Kconfig.

Derived from: nxp-imx/uboot-imx@93145335fbf5

Signed-off-by: Alifer Moraes <alifer.m@variscite.com>