LFU-571-1 imx: Fix ROMAPI driver to load container image
When using ROM API, the size and offset must be aligned with page size.
However when loading container image, the rule is not followed for
container header. It is ok for eMMC/SD/flexspinor, but fails for
flexspi NAND which page size is 4K.

Fix the issue by providing spl_romapi_read API to handle the unaligned
read and around container header size to page size

Signed-off-by: Ye Li <ye.li@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>