imx8mm_var_dart: Kconfig: drop sourcing freescale/common/Kconfig
The way that we use this file currently means that we have to guard it
in every platform Kconfig.  But it is also required in all NXP
platforms, including non-reference platforms.  Make all options in it
have appropriate dependencies so that we can include it a single time
under arch/Kconfig.

For further details, see:
nxp-imx/uboot-imx@93145335fbf5cbe7f6b1ca42263a156620a37f7f

Signed-off-by: Alexander Herfurtner <alexander.h@variscite.com>