imx93_var_som: Make dts selection account for CONFIG_DEFAULT_FDT_FILE
As of c6ccd1e, if fdt_file is unset, the default boot sequence will set
the device tree for the Symphony board whose version is dependent on the
current SoM version. Prior to this, if fdt_file was unset, it instead
relied on CONFIG_DEFAULT_FDT_FILE.

This new behavior invalidates any effect of CONFIG_DEFAULT_FDT_FILE.
To fix this, simply check whether CONFIG_DEFAULT_FDT_FILE is even
targeting a Symphony board, and if not, rely on the CONFIG, otherwise,
do the same and set the version accordingly.

Signed-off-by: Ken Sloat <ken.s@variscite.com>