imx8mp_var_dart: enable CONFIG_DM_SERIAL
Derived from commit 797cf4ec5210.

As per doc/develop/driver-model/migration.rst the deadline to
migrate to the config CONFIG_DM_SERIAL is set to U-Boot 2023.04.
NXP already migrated to that config in [1].

Currently SPL is not configured with DM, so we keep the bare metal
UART init and move it to SPL.

[1] nxp-imx/uboot-imx@ae188eb61c99

Signed-off-by: Alifer Moraes <alifer.m@variscite.com>