variscite: imx8mp_var_dart: refactor if condition for CONFIG_EXTCON_PTN5150
The struct extcon_ptn5150 usb_ptn5150 is defined through a #ifdef condition,
which is evaluated before compilation. This structure is used in the function
calling extcon_ptn5150_setup(), but this function is called in a regular C if
call, and it won't find the token usb_ptn5150 if CONFIG_EXTCON_PTN5150 is not
defined, leading to compiling error:

error: ‘usb_ptn5150’ undeclared (first use in this function)

Change the if condition that calls extcon_ptn5150_setup() function to #ifdef
expression.

Signed-off-by: Diego Dorta <diego.d@variscite.com>