LF-12177: misc: fsl_portals: add fixup for reserving QBMan private memory regions
Reserve DDR memory regions for QBMan base address registers (FBPR, FQD,
PFDR). Add the regions to the fdt passed to the kernel.

The memory regions are accessed in Linux and can be configured in BAR
registers only once per SoC reset. In order to guarantee the same memory
regions can be used by successive kernels without a reset (e.g. kexec),
reserve the regions from the bootloader so they persist between kernels.

There was a previous attempt to do this in Linux kernel commit
6ea4c0fe4570 ("soc/fsl/qbman: Update device tree with reserved memory"):
https://patchwork.kernel.org/project/linux-arm-kernel/patch/1564690599-29713-8-git-send-email-roy.pledge@nxp.com/

but that doesn't exactly do what's written on the tin. The changes made
through of_add_property() don't propagate to the flat device tree that
gets passed to the second kernel. They are just visible in the data
structures of the first kernel.

Fixup enabled for LS1046A SoCs.

Signed-off-by: Camelia Groza <camelia.groza@nxp.com>
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>