imx8mp_var_dart: reintroduce WCR initial configuration
Partially reverts commit b0f3dff, reintroducing the initial
configuration of the Watchdog Control Register (WCR).

In commit b0f3dff, set_wdog_reset() was removed from the board code
because its assignment was expected to be handled by the watchdog driver,
after setting the 'fsl,ext-reset-output' property in the device tree.

This commit aligns with NXP's latest changes. NXP removed set_wdog_reset()
in the commit nxp-imx/uboot-imx@a38ed61643a4, but was reintroduced it in
the commit nxp-imx/uboot-imx@504b56692135.

However, the WCR was not configured properly, and the WDT bit, which
determines whether the WDOG_B gets asserted upon a Watchdog timeout
event, was not set. This led to WDOG_B not being asserted after a timeout.
When WDOG_B is asserted (activated), it can trigger a broader system reset,
not just the processor itself. This is essential for systems where a simple
software reset of the processor is insufficient to recover from certain
failure states, and a full hardware reset (including peripherals and external
components) is necessary.

One of the issues caused was that the DDR could not be reinitialized
correctly after a watchdog-triggered reset when the board was in suspend mode,
causing a hang in U-Boot during DDR setting.

Signed-off-by: Alifer Moraes <alifer.m@variscite.com>