LF-11882-2 usb: xhci-imx8m: add workaround for USB3 host controller TXFIFO IC issue
On i.MX95(A) Soc, USB3 TXFIFO(RAM1) memory range 0x84800-0x88000 is
inaccessible, if write data to this memory the content is unchanged.
When the host controller fetch data from this memory range and transmit
the data to USB bus, data transfer error will occur. This issue happens
when the controller use default configuration of TXFIFO (automatically
calculated by HW). However, we can change the TXFIFO configuration by
GTXFIFOSIZE* to workaround this issue.

The host controller mainly uses TXFIFO for FS/LS, HS, SS and debug
capability cases. The size and offset is controlled by GTXFIFOSIZE0,
GTXFIFOSIZE1, GTXFIFOSIZE2 and GTXFIFOSIZE3 respectively.

The default value of GTXFIFOSIZE0/1 does not span over issue memory.
The default value of GTXFIFOSIZE2 is 0x02060811 and it means use memory
0x81030-0x850B8. With this change GTXFIFOSIZE2 will use memory 0x88000-
0x8C400.
The default value of GTXFIFOSIZE3 is 0A170022 and it means use memory
0x850B8-0x851C8. With this change GTXFIFOSIZE3 will use memory 0x81030-
0x81140.

This workaround is only for host mode.

Reviewed-by: Ye Li <ye.li@nxp.com>
Reviewed-by: Jun Li <jun.li@nxp.com>
Signed-off-by: Xu Yang <xu.yang_2@nxp.com>