imx8qxp_var_som defconfigs: disable HS400 mode
On a small percentage of eMMCs an error was observed when switching to
HS400 mode, when using this U-Boot branch. When working in HS200 mode,
the eMMCs always work fine.

Additional investigations, helped to identify the following community patch
varigit/uboot-imx@ef5ab0d13a7a mmc: do not send cmd13 if the parameter 'send_status' is 0 for __mmc_switch
which fixed the HS400 behavior in the rare cases the errors were
observed.

Apparently, the hang seems to be originate by a wrong sequence of
commands when switching to HS400 mode. Given the negligible draw-back
(30 ms slower while loading the kernel), to be conservative, simply
disable HS400 mode.

Signed-off-by: Pierluigi Passaro <pierluigi.p@variscite.com>