imx8mp_var_dart: set device tree based on SoM revision
With the introduction of the DART-MX8M-PLUS v2.x module, new device trees
have been added to support the updated hardware.

Read the module revision from the EEPROM and dynamically sets the appropriate
device tree based on the detected SoM version, ensuring that the correct device
tree is used for each module revision.

Signed-off-by: Alifer Moraes <alifer.m@variscite.com>