mx6ul_var_dart: Fix fdt_file logic to handle DTBs for VAR-SOM-6UL
This commit addresses the boot issue encountered on VAR-SOM-6UL
boards with Concerto and Symphony carrier boards when booting
from eMMC. The issue did not occur when booting from an SD card.
The logic was incorrectly appending the Wi-Fi extension to DTBs
that do not yet support this feature, leading to failures. This
fix ensures proper handling of the fdt_file for both legacy and
newer boards.

Signed-off-by: Diego Dorta <diego.d@variscite.com>