[RND-1518] variscite: imx8_eeprom: Add ddr fixup support
The EEPROM contains a diff that should be applied to the default DDR
timing configuration to support different DDR part numbers.

Sometimes NXP's DDR tool updates a DDR property that was already written to
the EEPROM. This function can be used to "fixup" known register / value
pairs to the correct value.

Signed-off-by: Nate Drude <nate.d@variscite.com>