common: imx8_eeprom: Mask is_spear flag when printing SOM revision
For i.MX8QM SOMs, bit 6 of the somrev field in the EEPROM is used as a
SPEAR-MX8 vs VAR-SOM-MX8 flag. This bit must not be taken into account
when decoding the SOM revision.

Previously, the SOM revision was calculated directly from somrev, so the
SPEAR flag would affect the major revision number. For example, a SPEAR-MX8
rev 1.4 was printed as "SOM revision: 3.4".

Use a local copy of somrev and mask out bit 6 when printing the revision
for CONFIG_TARGET_IMX8QM_VAR_SOM, so the board ID logic still uses the
flag but the printed revision matches the actual SOM.

Signed-off-by: Diego Dorta <diego.d@variscite.com>